refactor(archive-preview): extract preview rendering and zip manifest parsing
